Factor out createPrompt into own file
