+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/use-suggestions-from-openai.js
@@ -3,11 +3,11 @@ import { useSelect } from '@wordpress/data';
 import { useEffect, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import MarkdownIt from 'markdown-it';
+import { createPrompt } from './create-prompt';
 
 const useSuggestionsFromOpenAI = ( {
 	clientId,
 	content,
-	createPrompt,
 	setAttributes,
 	setErrorMessage,
 	tracks,
